new structure of pr folder. Edited Makefile
