+++ pdp-11/pr/Makefile
@@ -1,10 +1,34 @@
+IDIR=include
 CC=gcc
-CFLAGS=-I -g.
-DEPS = pdp.h
-OBJ = pdp_main.o pdp_main-func.o pdp_add-func.o pdp_tests.o pdp_run.o pdp_do-func.o
+CFLAGS=-I$(IDIR) -g
+BIN=pdp11.exe
 
-%.o: %.c $(DEPS)
+
+SDIR=src
+ODIR=obj
+
+
+_DEPS = pdp.h
+DEPS = $(patsubst %,$(IDIR)/%,$(_DEPS))
+
+
+_OBJ = pdp_main.o pdp_main-func.o pdp_add-func.o pdp_tests.o pdp_run.o pdp_do-func.o
+OBJ = $(patsubst %,$(ODIR)/%,$(_OBJ))
+
+
+all: $(BIN)
+
+
+$(ODIR)/%.o: $(SDIR)/%.c $(DEPS)
 	$(CC) -c -o $@ $< $(CFLAGS)
 
-pdp11.exe: $(OBJ)
-	$(CC) -o $@ $^ $(CFLAGS)
+
+$(BIN): $(OBJ)
+	$(CC) -o $@ $^ $(CFLAGS) $(LIBS)
+
+
+.PHONY: clean
+
+
+clean:
+	rm -f $(ODIR)/*.o $(SDIR)/*~ $(INCDIR)/*~
