--- pdp-11/pr/test.o
@@ -1,26 +1,0 @@
-0040 0005
-01
-20
-00
-ff
-ba
-0042 0003
-6b
-1c
-0f
-0200 000f
-c6
-15
-00
-02
-c1
-15
-3d
-8e
-f7
-09
-02
-00
-00
-00
-66
