--- pdp-11/pr/test_arg.o
@@ -1,4 +1,0 @@
-0042 0003
-6b
-1c
-0f
